refactor(docs): dropdown dx
