+++ apps/docs/content/components/dropdown/disabled-keys.ts
@@ -1,26 +1,4 @@
-const App = `import {Dropdown, DropdownTrigger, DropdownMenu, DropdownItem, Button} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Dropdown >
-      <DropdownTrigger>
-        <Button 
-          variant="bordered" 
-        >
-          Open Menu
-        </Button>
-      </DropdownTrigger>
-      <DropdownMenu aria-label="Example with disabled actions" disabledKeys={["edit", "delete"]}>
-        <DropdownItem key="new">New file</DropdownItem>
-        <DropdownItem key="copy">Copy link</DropdownItem>
-        <DropdownItem key="edit">Edit file</DropdownItem>
-        <DropdownItem key="delete" className="text-danger" color="danger">
-          Delete file
-        </DropdownItem>
-      </DropdownMenu>
-    </Dropdown>
-  );
-}`;
+import App from "./disabled-keys.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
