+++ apps/docs/content/components/dropdown/dynamic.raw.jsx
@@ -1,0 +1,41 @@
+import {Dropdown, DropdownTrigger, DropdownMenu, DropdownItem, Button} from "@nextui-org/react";
+
+export default function App() {
+  const items = [
+    {
+      key: "new",
+      label: "New file",
+    },
+    {
+      key: "copy",
+      label: "Copy link",
+    },
+    {
+      key: "edit",
+      label: "Edit file",
+    },
+    {
+      key: "delete",
+      label: "Delete file",
+    },
+  ];
+
+  return (
+    <Dropdown>
+      <DropdownTrigger>
+        <Button variant="bordered">Open Menu</Button>
+      </DropdownTrigger>
+      <DropdownMenu aria-label="Dynamic Actions" items={items}>
+        {(item) => (
+          <DropdownItem
+            key={item.key}
+            className={item.key === "delete" ? "text-danger" : ""}
+            color={item.key === "delete" ? "danger" : "default"}
+          >
+            {item.label}
+          </DropdownItem>
+        )}
+      </DropdownMenu>
+    </Dropdown>
+  );
+}
